Cache patched module dependencies
A lot more work would be needed to make this approach viable.
Caching dependencies has the advantage that repeat calls to the context
manager yield the same sys.modules state, so that different parts of
a module can be independently patched. On the other hand, permitting
nested patching contexts for the same module would allow patching an
entire package by nesting calls in dependency order, which may be
preferable.